+++ diive/__init__.py
@@ -47,6 +47,8 @@ from diive.pkgs.gapfilling.xgboost_ts import XGBoostTS as xgboost_ts
 from diive.pkgs.gapfilling.xgboost_ts import XGBoostTS
 from diive.pkgs.gapfilling.mds import FluxMDS as flux_mds
 from diive.pkgs.gapfilling.mds import FluxMDS
+from diive.pkgs.binary.extract import get_encoded_value_from_int as get_encoded_value_from_int
+from diive.pkgs.binary.extract import get_encoded_value_series as get_encoded_value_series
 
 __all__ = [
     # Configs
@@ -115,4 +117,8 @@ __all__ = [
     'XGBoostTS',
     'flux_mds',
     'FluxMDS',
+
+    # Packages: Binary
+    'get_encoded_value_from_int',
+    'get_encoded_value_series',
 ]